requirements.txt
attrs==20.2.0
certifi==2020.6.20
dotmap==1.3.17
iniconfig==1.0.1
more-itertools==8.5.0
packaging==20.4
pika==1.1.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.2
pytest-env==0.6.2
python-dotenv==0.14.0
PyYAML==5.3.1
six==1.15.0
toml==0.10.1
